View tests

+++ classes/SQL.php
@@ -38,23 +38,30 @@ class SQL {
 
   public static function & connection(){
     if(null === static::$pdo) {
-      static::$pdo = new PDO(
-          static::$connection['dsn'],
-          static::$connection['username'],
-          static::$connection['password'],
-          static::$connection['options']
+      try {
+        static::$pdo = new PDO(
+            static::$connection['dsn'],
+            static::$connection['username'],
+            static::$connection['password'],
+            static::$connection['options']
 
-      );
-      Event::triggerOnce('core.sql.connect');
+        );
+      } catch(Exception $e) {
+        static::$pdo = null;
+      }
+      Event::triggerOnce('core.sql.connect',static::$connection,static::$pdo);
     }
     return static::$pdo;
   }
 
   public static function prepare($query){
+    if(!static::connection()) return false;
     return isset(static::$queries[$query]) ? static::$queries[$query] : (static::$queries[$query] = static::connection()->prepare($query));
   }
 
   public static function exec($query, $params=[]){
+    if(!static::connection()) return false;
+    
     if (false==is_array($params)) $params = (array)$params;
     $query = Filter::with('core.sql.query',$query);
     if($statement = static::prepare($query)){
@@ -82,11 +89,15 @@ class SQL {
   }
 
   public static function value($query, $params=[], $column=0){
+    if(!static::connection()) return false;
+    
     $res = static::exec($query,$params);
     return $res ? $res->fetchColumn($column) : null;
   }
 
   public static function each($query, $params=[], callable $looper = null){
+    if(!static::connection()) return false;
+    
     // ($query,$looper) shorthand
     if ($looper===null && is_callable($params)) {$looper = $params; $params = [];}
     if( $res = static::exec($query,$params) ){
@@ -98,6 +109,8 @@ class SQL {
   }
 
   public static function single($query, $params=[], callable $handler = null){
+    if(!static::connection()) return false;
+    
     // ($query,$handler) shorthand
     if ($handler===null && is_callable($params)) {$handler = $params; $params = [];}
     if( $res = static::exec($query,$params) ){
@@ -109,6 +122,8 @@ class SQL {
   }
 
  public static function run($script){
+    if(!static::connection()) return false;
+    
     $sql_path = Options::get('database.sql.path',APP_DIR.'/sql');
     $sql_sep  = Options::get('database.sql.separator',';');
     if (is_file($f = "$sql_path/$script.sql")){
@@ -121,10 +136,14 @@ class SQL {
   }
 
   public static function all($query, $params=[]){
-    return static::each($query,$params);
+   if(!static::connection()) return false;
+  
+   return static::each($query,$params);
   }
 
   public static function delete($table, $pks=null, $pk='id', $inclusive=true){
+    if(!static::connection()) return false;
+    
     if (null===$pks) {
       return static::exec("DELETE FROM `$table`");
     } else {
@@ -135,6 +154,8 @@ class SQL {
   }
 
   public static function insert($table, $data=[]){
+    if(!static::connection()) return false;
+    
     if (false==is_array($data)) $data = (array)$data;
     $k = array_keys($data);
     asort($k);
@@ -146,6 +167,8 @@ class SQL {
   }
 
   public static function update($table, $data=[], $pk='id', $extra_where=''){
+    if(!static::connection()) return false;
+    
     if (false==is_array($data)) $data = (array)$data;
     if (empty($data[$pk])) return false;
     $k = array_keys($data);
@@ -157,6 +180,8 @@ class SQL {
   }
 
   public static function insertOrUpdate($table, $data=[], $pk='id', $extra_where=''){
+    if(!static::connection()) return false;
+    
     if (false==is_array($data)) $data = (array)$data;
     if (empty($data[$pk])) return static::insert($table, $data);
     if( (string) static::value("SELECT `$pk` FROM `$table` WHERE `$pk`=? LIMIT 1", [$data[$pk]]) === (string) $data[$pk] ){
